+++ CHANGELOG.md
@@ -6,6 +6,9 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+### Changed
+ - Re-initializing the Windows audio device if there is a COMException
+
 ## [10.1.0] - 2021-10-04
 ### Changed
  - Occupancy sensors implement new IOccupancySensorControl SupportedFeatures and PeopleCount